requirements.txt
--index-url https://download.pytorch.org/whl/cu121

torch
torchvision
torchaudio
scikit-learn
tqdm
PyYAML